backend/pyproject.toml
[project]
name = "nemori-backend"
version = "1.0.0"
description = "Nemori AI Memory Assistant - Backend Service"
authors = [{ name = "FellouAI" }]
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    # AI/LLM
    "openai>=1.60.0",
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.0",
    "langgraph>=0.5.0",
    # Vector store
    "chromadb>=1.0.0",
    # Image processing
    "pillow>=11.0.0",
    "imagehash>=4.3.1",
    "mss>=9.0.1",
    # Async/IO
    "aiofiles>=24.0.0",
    "websockets>=14.0",
    "httpx>=0.28.0",
    # Utils
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.0.0",
    "isort>=6.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = ["*.pyc", "__pycache__", "*.egg-info"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
